Narrow super tenants GET import boundary
